Adding "datetime"
On second thought, perhaps it's better to give the user (developer) options to opt in to date, time, or datetime, as desired. I've added the option, and removed the "^" and "$" start and end markers appropriately. This could possibly do it only if it's "datetime" being specified, but it depends on whether there's a risk not having it?