[.NET Timexlib] Implement missing features of TimexProperty.ToString()
This implements the missing features of TimexProperty.ToString(), adding support for date ranges, date + duration, etc.

This also fixes a bunch of off-by-one bugs when adding days or months to a Timex where it would not take into account potential month/year overflows.

It also adds caching of the `Types` hashset of the Timex, so that it doesn't recalculate it every time the property is accessed. Instead, `Types` get cleared whenever a meaningful piece of data in the `TimexProperty` is changed.

Some tests have been converted to using `DataRow` instead of inline asserts, so that the test log will show the data being tested. It should also make it easier to see what the tests are doing.

There's a lot of static helper methods that have been made into extension methods, because they're already in static classes so there's no overhead in doing it, and it makes caller code shorter and hopefully simpler to read.